property-manager-product-migration/requirements.txt
# requirements.txt
requests
pandas
akamai-edgegrid
